requirements.txt
qdrant-client
sentence-transformers
llama-index
PyMuPDF
python-dotenv
tqdm

